Final Version JS complete
The main aim of this project is to understand when to use a function
factory and the reveal module layout.

function factory: to create objects with inheritance of functions
reveal module layout: to create private variables

JS is complete but the CSS is clearly lacking. I am here to learn JS,
so that's all good!